requirements.txt
# Núcleo científico
numpy>=1.26,<2
pandas>=2.1,<2.3
scikit-learn>=1.3,<1.6
joblib>=1.3,<2
pyyaml>=6.0,<7
matplotlib>=3.7,<3.9

# Tracking
mlflow>=2.9,<3

# Utilidades
requests>=2.31,<3

